Use IPv4 to talk to www.tumblr.com
The privacy consent form doesn't actually do anything if it thinks
it's being accessed from outside the EU. This is a problem if you
have an IPv6 tunnel which terminates outside of the EU, because
while individual Tumblr blog pages are only served over IPv4,
www.tumblr.com itself also listens to IPv6.

Make cURL only use IPv4 in its call to get the cookie, to make sure
we use the same routing path to www.tumblr.com as we do to the
individual blog page.